refactor(brain): make the doctor a registry of independent checks
`doctor.ts` was 1819 lines holding twenty-six unrelated diagnostics in
one body (cognitive complexity 290, MI -63.2). The checks share nothing
but their inputs, which is what made the single file a Single
Responsibility violation rather than merely a long one.

Each check family now lives in a module named for what it diagnoses, and
`doctor.ts` keeps its path as the registry that runs them: a 26-entry
array a new check joins by being appended to. Every check declares its
own fail-soft flag beside its implementation, which encodes exactly the
guard placement the old body had - structural checks throw, lints are
swallowed so one broken scan cannot mask the rest.

The finding contract is intact: 41 diagnostic codes before, 41 after,
verified by running the census's own regexes over both trees. The census
ratchet is strengthened rather than weakened - it now enumerates the
check modules from disk instead of reading one file, with a vacuity
assertion so a renamed directory cannot silently shrink the population.
`classifyParseError` moved verbatim and still reads the typed error's
fields; no message-string parsing was reintroduced.

Four defects fixed: a `void parseFrontmatter;` statement whose comment
claimed it prevented dependency stripping while the import was genuinely
used all along, a code literal written raw in one branch and as its
constant in the other, a day-in-milliseconds expression repeated in two
cutoffs, and a config parameter re-narrowed from a value already on the
context.

One ordering change, stated rather than hidden: the shared context now
resolves up front instead of interleaving between checks. All of it is
pure reads, so no finding moves; the only difference is that an unpinned
clock is read a few milliseconds earlier, and every age-based lint has
day granularity.
